% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_aggregate_metabarlist.R
\name{aggregate_metabarlist}
\alias{aggregate_metabarlist}
\title{Aggregate PCR replicates}
\usage{
aggregate_metabarlist(metabarlist, groups = "sample_id",
  method = "sum")
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}

\item{groups}{a column name in the `pcrs`table corresponding to the sample names to which
pcr replicates belongs. Default is the `sample_id` column of the table `pcrs` from the
\code{\link{metabarlist}} object.}

\item{method}{a method of aggregation returning a distance matrix. Can be \code{"sum"}, \code{"mean"},
\code{"prob"}. Default is "sum"}
}
\value{
A \code{metabarlist}
}
\description{
Aggregate PCR replicates in a \code{metabarlist} object.
}
\details{
The function \code{aggregate_metabarlist} is typically used at the end of the data filtration process and aims at aggregating reads and the pcr related information at the sample level. Different methods of aggregation is available:

#'\itemize{
\item{With method \code{"sum"}, reads of pcr replicates are summed for each OTU}
\item{With method \code{"mean"}, reads of pcr replicates are averaged for each OTU. Results are rounded so that to obtain genuine count data}
\item{With method \code{"prob"}, the probability of detection is returned for each OTU}

For all methods, the `pcr` table is simplified.

## LZ: Fred's coa method to include too?
}

##TODO think about other methods of aggregation: e.g. weighted mean, sum and rarefaction, rarefaciton and sum, what else? or let it as a free function (but then this is contradictory with the idea of "simplifying stuff")
##TODO think about how to simplify the pcr table. Wonder if it should not be removed...
}
\examples{

data(soil_euk)

soil_euk_ag <- aggregate_metabarlist(soil_euk, method="sum")
summary_metabarlist(soil_euk)
summary_metabarlist(soil_euk_ag)

soil_euk_ag <- aggregate_metabarlist(soil_euk, method="prob")
summary_metabarlist(soil_euk)
summary_metabarlist(soil_euk_ag) ## output stat are surprising

}
\author{
Frédéric Boyer, Lucie Zinger
}
