% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_motus.R
\name{aggregate_motus}
\alias{aggregate_motus}
\title{Aggregate MOTUs}
\usage{
aggregate_motus(metabarlist, groups = NULL, FUN = FUN_agg_motus_sum)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}

\item{groups}{a grouping vector or factor for MOTUs.
\code{NA} values treated as a group level}

\item{FUN}{a function of MOTU aggregation.
Default is the sum of MOTUs reads abundances for each grouping value.}
}
\value{
A \code{metabarlist} where the table `reads` contains MOTUs abundances aggregated according to a grouping vector/factor (e.g. taxonomic assignment at the phylum level), using a method defined in \code{FUN} and where number of columns of tables `reads` will be equal to that of the number of groups in `groups`.
}
\description{
Aggregate MOTUs in a \code{metabarlist} object according to a grouping factor or vector
}
\details{
The function \code{aggregate_motus} is typically used for aggregating MOTUs at a given taxonomic resolution. The user is free to use its own method of aggregation, but one method is often used and therefore pre-encoded:

#'\itemize{
\item{\code{"FUN_agg_motus_sum"}: reads of MOTUs in a given group are summed for each pcr.}
}

After aggregation, the information contained in the `motus` table corresponds to the
information of the most abundant MOTU in a given group.
}
\examples{

data(soil_euk)

## With MOTU phylum assignment as grouping factor and
## default grouping aggregation (sum reads across replicates)
soil_euk_ag <- aggregate_motus(soil_euk, groups = soil_euk$motus$phylum_name)
summary_metabarlist(soil_euk)
summary_metabarlist(soil_euk_ag)

## With a custom function (here equivalent to FUN_agg_sum,
## i.e. summing all MOTUs abundance across groups)
soil_euk_ag <- aggregate_motus(soil_euk, groups = soil_euk$motus$phylum_name,
                               FUN = function(metabarlist, groups){
                                       t(rowsum(t(metabarlist$reads), groups))})

}
\seealso{
@seealso \code{\link{aggregate_pcrs}}, \code{\link{apply}}, \code{\link{aggregate}}
}
\author{
Lucie Zinger
}
