% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_aggregate_reads.R
\name{aggregate_reads}
\alias{aggregate_reads}
\title{Check that a list of tables is in the correct format as a metabarlist}
\usage{
aggregate_reads(metabarlist, groups = metabarlist$pcrs$sample_id,
  FUN = sum)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}

\item{groups}{a vector containing the identifier of replicate. The vector must has the same length of the table `PCRs` from a \code{\link{metabarlist}} object. Default = metabarlist$pcrs$sample_id}

\item{FUN}{a function returning a distance matrix. The distance matrix should be a object `dist` which has the same length of table `reads`.}
}
\value{
TRUE or throws a stop
}
\description{
Test that a list of tables containing information on MOTU abundances, MOTU characteristics, pcr characteristics, and sample characteristics form a congruent \code{\link{metabarlist}}
}
\details{
#'
}
\examples{

data(soil_euk)

aggregate_reads(soil_euk)

}
\author{
Frédéric Boyer
}
