% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_check_metabarlist.R
\name{check_metabarlist}
\alias{check_metabarlist}
\title{Test that the provided list is a well formed metabaRffe list}
\usage{
check_metabarlist(metabarlist)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}
}
\value{
TRUE or throws an stop
}
\description{
Test that the provided list is a well formed metabaRffe list
}
\details{
Check for the properties awaited for a well formed metabaRffe list:
\itemize{
\item {is a list with three attributes named `reads`, `motus` and `pcrs`}
\item {`reads` is a numeric matrix}
\item {`motus`, `pcrs` and `samples` are of type `data.frame`}
\item {`reads` have the exact same row names and columns as rownames of `motus` and `pcrs` (same order also)}
\item {`pcrs` have mandatory columns, i.e.  `sample_id`, `type` and `control_type`}
\item {values in `type` are properly defined, i.e. `sample` or `control`}
\item {values in and `Control_type` are properly defined, i.e. `sequencing`, `pcr`, `extraction`, `positive`}


\item {issue warnings if}
\item {`motus` and `pcrs` have not mandatory columns for design description, i.e. tag_fwd, tag_rev, primer_fwd, primer_rev, plate_no, plate_col and plate_row for `pcrs`
  and sequence for `motus`}
}

The function issues a stop if these informations are not well recorded

Moreover, the function issues a warning if any sample or OTU is associated to a count of 0
}
\examples{

data(soil_euk)

check_metabarlist(soil_euk)

}
\author{
Clément Lionnet & Frédéric Boyer
}
