% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_extract_table.R
\name{extract_table}
\alias{extract_table}
\title{Extract tables from a \code{metabarlist} object}
\usage{
extract_table(metabarlist, table = extract_table_methods)
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object}

\item{table}{the table required for extraction.
Can be one of `reads`, `motus`, `pcrs`, or `samples`}
}
\value{
a numeric matrix `reads` or a data frame `motus`, `pcrs`, or `samples`
}
\description{
\code{extract_table} is a generic function which extracts a table from a \code{\link{metabarlist}} object.
}
\details{
All \code{metabarlist} objects are composed of four different tables that can be extracted as single tables with this function.
}
\examples{

data(soil_euk)

reads <- extract_table(soil_euk, "reads")
motus <- extract_table(soil_euk, "motus")

all(colnames(reads) == rownames(motus))
is.matrix(reads)
is.data.frame(motus)

}
\seealso{
\code{\link{metabarlist_generator}}
}
\author{
Lucie Zinger
}
