% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_fasta_generator.R
\name{fasta_generator}
\alias{fasta_generator}
\title{Generate a fasta file on a subset of sequences}
\usage{
fasta_generator(metabarlist, id = rownames(metabarlist$motus),
  output_file = NULL, annotation = NULL, annot_sep = " ")
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object (Required)}

\item{id}{a character vector containing identifiers for the OTUs of interest. Default: rownames(metabarlist$motus)}

\item{output_file}{the path/name of the output file. (Required)}

\item{annotation}{a dataframe containing the additional information adding in the header. Each row name is an OTU identifiers and each column header is the information key (correspond to "key=" in the fasta header). Default: NULL}

\item{annot_sep}{the annotation separator. Default: " "}
}
\value{
a fasta file
}
\description{
Generate a fasta file on a subset of OTUs from a \code{\link{TODEFINE}} object.
}
\details{
Creates a fasta file from a set of sequences of interest. Obviously, \code{id} and \code{seq} vectors should be in agreement. The header can easily be amended with further information with e.g. \code{paste}.
}
\examples{

data(soil_euk)

# export in fasta the 10 most abundant MOTUs
idx <- order(soil_euk$motus$count, decreasing = T)[1:10]
fasta_generator(
  soil_euk, rownames(soil_euk$motus)[idx],
  "Dominants.fasta"
)

# export in fasta the 10 most abundant MOTUs and their abundance
idx <- order(soil_euk$motus$count, decreasing = T)[1:10]
annotation <- data.frame(
  abundance = soil_euk$motus$count[idx],
  row.names = rownames(soil_euk$motus)[idx]
)
fasta_generator(
  soil_euk, rownames(soil_euk$motus)[idx],
  "Dominants.fasta",
  annotation
)
}
\author{
Lucie Zinger
}
