% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_ggplate.R
\name{ggplate}
\alias{ggplate}
\title{Plotting an attribute in the PCR plate design}
\usage{
ggplate(x, attr, coords, type)
}
\arguments{
\item{x}{a \code{\link{TODEFINE}} object}
}
\value{
a ggplot
}
\description{
Plots an attribute of amplicons from a \code{\link{TODEFINE}} object in the correspondint PCR plate design
}
\details{
To write
}
\examples{

data(soil_euk)


}
\author{
Fred Boyer, Lucie Zinger
}
