% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_hill_rarefaction.R
\name{hill_rarefaction}
\alias{hill_rarefaction}
\title{Generate rarefaction curves using the hill numbers}
\usage{
hill_rarefaction(x, boot = 10)
}
\arguments{
\item{x}{a community matrix from a \code{\link{TODEFINE}}}

\item{boot}{nb of resampling to estimate Dq at a given sequencing depth}
}
\value{
a table of data rarefied at different sequencing depth
}
\description{
Generate rarefaction curves from a \code{\link{TODEFINE}} object using hill numbers diversity indices, as well as the Good's coverage index.
}
\details{
Build a rarefaction analysis of amplicons in a \code{\link{TODEFINE}} object using Hill numbers for q={0,1,2}. These indices are equivalents to richness (q=0), exponential Shannon (q->1), and inverse simpson (q=2). The function also returns the Good's coverage index. But note that this index should be interpreted carefully, as it is based on singletons in each amplicons, some of them being filtered during the bioinformatic process (filtering of absolute singeltons).
}
\examples{

data(soil_euk)

#create subset
idx = c(grep("-J", soil_euk$pcr$point_id), which(is.na(soil_euk$pcr$Control_type)==F))
soil_euk_raref.s = hill_rarefaction(soil_euk$reads[idx,], boot=10)

#create a vector differenciating samples of different type matching with \\code{hill_rarefaction} output
soil_euk$pcr$sample_type = as.vector(soil_euk$pcr$Control_type)
soil_euk$pcr$sample_type[is.na(soil_euk$pcr$sample_type)] = "Biological samples"
vectype = soil_euk$pcr$sample_type[match(soil_euk_raref.s$sample, rownames(soil_euk$pcr))]

gghill_rarefaction(soil_euk_raref.s, vectype)

}
\references{
Chao, A., Chiu, C. H., & Jost, L. (2014). Unifying species diversity, phylogenetic diversity, functional diversity, and related similarity and differentiation measures through Hill numbers. Annual review of ecology, evolution, and systematics, 45, 297-324.
}
\author{
Lucie Zinger
}
