% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_metabarlist_generator.R
\name{metabarlist_generator}
\alias{metabarlist_generator}
\title{Import data to create a metabarlist object}
\usage{
metabarlist_generator(file_reads, file_motus, file_pcrs, file_samples,
  files_sep = "\\t")
}
\arguments{
\item{file_reads}{the path for the MOTU abundance table. Rows of the table should correspond to pcrs, and the columns to MOTUs. The first column of this table should correspond to the names of the pcrs.}

\item{file_motus}{the path for the MOTU characteristics table (e.g. taxonomy, sequence, etc.). Rows of the table should correspond to MOTUs, and the columns to their characteristics. The first column of this table should contain MOTUs names. Mandatory fields: `sequence`, i.e. the most abundant sequence of the MOTU.}

\item{file_pcrs}{the path for the pcrs characteristics table (e.g. tags, primers, plate wells, etc.). Mandatory fields: (i) `sample_id`, i.e. the name of each sample. (ii) `type`, i.e. the type of pcr; can be `sample` or `control`. (iii) `control_type`, i.e. the type of control if applicable. Should be: `NA` for samples, `extraction` for extraction negative controls, `pcr` for pcr negative controls, `sequencing` for sequencing negative controls (e.g. unused tag combinations), and `positive` for positive controls.}

\item{file_samples}{the path for the sample characteristics table. The first column of this table shuld contain the sample names.}

\item{files_sep}{separator used to read the different tables. Should be the same in all tables. Default is tabulation.}
}
\value{
a \code{\link{metabarlist}} object
}
\description{
Importing and formatting files to create a \code{\link{metabarlist}} object.
}
\details{
This function aims at importing data into \R to create a \code{\link{metabarlist}} object. The four files required are imported in \R, included into a list of class \code{\link{metabarlist}}, and congruencies between all tables are tested with the \code{\link{check_metabarlist}} function.
}
\examples{

\dontrun{
soil_euk = metabarlist_generator(file_reads = "data-raw/litiere_euk_reads.txt",
                                      file_motus = "data-raw/litiere_euk_motus.txt",
                                      file_pcrs = "data-raw/litiere_euk_pcrs.txt",
                                      file_samples = "data-raw/litiere_euk_samples.txt")
}

}
\seealso{
\code{\link{check_metabarlist}}
}
\author{
Lucie Zinger & Clément Lionnet & Frédéric boyer
}
