% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabarlist_to_phyloseq.R
\name{metabarlist_to_phyloseq}
\alias{metabarlist_to_phyloseq}
\title{Converts a metabarlist object in a phyloseq object}
\usage{
metabarlist_to_phyloseq(metabarlist, tax_colnames, include_controls = FALSE)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object aggregated at sample level}

\item{tax_colnames}{a character vector with the column names to keep from the \code{motus} table of the \code{metabarlist} object}

\item{include_controls}{logical: if \code{TRUE}, controls are added in the \code{phyloseq} \code{sample_data} object, as well as the columns of the \code{pcrs} table. MOTUs abundance in controls should be present in the \code{motus} table The default (\code{FALSE}) is not to include controls. In that case, the columns of the \code{pcrs} table are added to the \code{sample_data} object and the \code{motus} table should not include abundances in controls.}
}
\value{
A \code{phyloseq} object with three components: \code{otu_table}, \code{sample_data} and \code{taxonomy_table}
}
\description{
Converts a \code{metabarlist} object in a \code{phyloseq} object for downstream diversity analyses.
}
\details{
The function \code{metabarlist_to_phyloseq} converts a \code{metabarlist} object into a \code{phyloseq} object. A \code{phyloseq} object can contain several component data, such as \code{otu_table}, \code{sample_data} and \code{taxonomy_table}. The \code{otu_table} is equivalent to the \code{reads} matrix, the \code{sample_data} is equivalent to the \code{samples} data frame and the \code{taxonomy_table} corresponds to the taxonomic informations that may be contained in the \code{motus} data frame. Information from the \code{pcrs} table is included in the \code{sample_data} component.
}
\examples{

data(soil_euk) # load soil_euk data
soil_euk_sub <- subset_metabarlist(soil_euk, "pcrs", soil_euk$pcrs$type == "sample")
soil_euk_agg <- aggregate_pcrs(soil_euk_sub)
tax_colnames <- c("phylum_name", "class_name", "order_name", "family_name", "genus_name",
 "species_name")
soil_euk_physeq <- metabarlist_to_phyloseq(soil_euk_agg, tax_colnames)
soil_euk_physeq

soil_euk_agg <- aggregate_pcrs(soil_euk)
soil_euk_physeq <- metabarlist_to_phyloseq(soil_euk_agg, tax_colnames,
 include_controls = TRUE)
soil_euk_physeq

}
\seealso{
the \code{phyloseq} package
}
\author{
Anne-Sophie Benoiston
}
