% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ngsfilter.R
\name{read_ngsfilter}
\alias{read_ngsfilter}
\title{Import an ngsfilter file}
\usage{
read_ngsfilter(file, additional.sep = "=", ...)
}
\arguments{
\item{file}{path/name of the input ngsfilter txt file}

\item{additional.sep}{character indicating the field separator in the additional info column in
the ngsfilter filter file.}

\item{...}{other arguments to be pasted from \code{read.table}}
}
\value{
a data.frame
}
\description{
Import an ngsfilter file formatted for the \code{OBITools} package (Boyer et al. 2016).
}
\details{
The function \code{read_ngsfilter} creates a `pcrs` table from an ngsfilter file formatted for the \code{OBITools} package (Boyer et al. 2016).
}
\examples{

\donttest{

dir <- tempdir()
url = "https://raw.githubusercontent.com/metabaRfactory/metabaR_external_data/master/"

ngsfilter_file = "ngsfilter_GWM-768.new_2.txt"
ngsfilter_url = paste(url, ngsfilter_file, sep="")
ngsfilter_path <- file.path(dir, ngsfilter_file)
download.file(ngsfilter_url, ngsfilter_path)

ngsfilter <- read_ngsfilter(
              file = ngsfilter_path,
              sep = "\t",
              additional.sep = "=")

}

}
\references{
Boyer, F., Mercier, C., Bonin, A., Le Bras, Y., Taberlet, P., & Coissac, E. (2016). obitools: a unix-inspired software package for DNA metabarcoding. Molecular Ecology Resources, 16(1), 176-182.
}
\seealso{
\code{\link{obifiles_to_metabarlist}}
}
\author{
Lucie Zinger
}
