% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_silva_annotator.R
\name{silva_annotator}
\alias{silva_annotator}
\title{Integrating SILVANgs pipeline taxonomic annotations}
\usage{
silva_annotator(motus, silva.path, clust.path)
}
\arguments{
\item{silva.path}{path to a table from the SILVANgs pipeline, typically zipfile>ssu>exports>xxx---ssu---otus.csv}

\item{clust.path}{path to a file from the SilvaNgs pipeline indicating otu cluster membership zipfile>ssu>stats>sequence_cluster_map>data>xxx---ssu---sequence_cluster_map---tmptaxo.clstr}

\item{x}{a motus dataframe from a \code{\link{TODEFINE}} object}
}
\value{
a motus dataframe for a \code{\link{TODEFINE}} object including the taxonomic assignemnts from silva
}
\description{
Importing and formatting SILVANgs pipeline taxonomic annotation for a \code{\link{TODEFINE}} object.
}
\details{
To amend
}
\examples{

}
\author{
Lucie Zinger
}
