% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_soil_euk.R
\docType{data}
\name{soil_euk}
\alias{soil_euk}
\title{Soil eukaryote DNA metabarcoding data from French Guiana}
\format{
An object of class \code{metabarlist}; see \code{\link[metabaRffe]{metabarlist}}.
}
\source{
\href{LINK WEB TO DEFINE}{DNA metabarcoding data archive} for obitools ouput data and contextual data.
}
\usage{
data(soil_euk)
}
\description{
Data from a DNA metabarcoding experiment assessing the diversity of soil eukaryotes in two tropical forests in French Guiana.
}
\details{
Samples were collected at 2 sample sites in contrasting habitats:
\itemize{
  \item{}{The Mana site located in a white sand forest, characterized highly oligotrophic soils and tree species adaptated to the local harsh conditions.}
  \item{}{The Petit Plateau site is located in pristine rainforest (Nouragues natural reserve) characterized by soils richer in clay and organic matter.}
 }
At each site, 16 samples (mesh grid size of 20 m) were collected in from a 1 ha plot.
For each sampling point, two types of substrate were collected:
\itemize{
  \item{}{Soil: composite sample of 5 soil cores}
  \item{}{Litter: sampling of 1 m2 of surface leaf litter}
  \item{}{=>A total of 64 DNA extracts + 4 extraction blank controls (1 per site and compartment)}
}
For each DNA extract, a short region of the 18S rRNA (Taberlet et al. 2018) was
amplfied by PCR in quadruplicate, following the protocol described in Zinger et al. 2019.
The resulting amplicons were pooled and sequenced on an Illumina HiSeq platform, using
the paired-end technology.

The total experiment hence includes the following amplicon types:
\itemize{
  \item{}{4 replicates / samples (n = 256)}
  \item{}{4 replicates for the 4 extraction blank controls (n = 16)}
  \item{}{4 replicates for 8 PCR blank controls (n = 32)}
  \item{}{4 replicates for 12 sequencing blank controls (n = 48)}
  \item{}{4 replicates for 8 positive controls (plant DNA from 16 species; n = 32)}
  \item{}{=> 384 amplicons in total}
}

The retrieved data were then processed using the OBITools (Boyer et al. 2016) and SUMACLUST (Mercier et al. 2013) packages. Briefly, paired-end reads were assembled, assigned to their respective samples/marker and dereplicated. Low-quality sequences (containing Ns, shorter than 50 bp or singletons) were excluded; the remaining sequences were clustered into operational taxonomic units (OTUs) using SUMACLUST at a sequence similarity threshold of 0.97. The representative sequence of each OTU (the most abundant one) was assigned a taxonomic clade using a databased built from the EMBL (release 136) with the ecoPCR program (Ficetola et al., 2010).


The data `soil_euk` is a \code{metabarlist} containing four tables

\itemize{
\item{}{`reads`: a numeric matrix containing MOTU abundances (expressed as a number of reads) for each pcr (i.e. technical replicates of both biological samples and positive and negative controls)}
\item{}{`motus`: a dataframe containing MOTU characteristics (e.g. taxonomy, sequence) for each MOTU)}
\item{}{`pcrs`: a dataframe containing information on each pcr (e.g. control type, pcr wells, etc.)}
\item{}{`samples`: a dataframe containing information on each environmental sample (e.g. habitat type, etc.)}
}
}
\examples{
data(soil_euk)
summary_metabarlist(soil_euk)
}
\references{
Boyer, F., Mercier, C., Bonin, A., Le Bras, Y., Taberlet, P., & Coissac, E. (2016). obitools: a unix‐inspired software package for DNA metabarcoding. Molecular Ecology Resources, 16(1), 176-182.

Ficetola, G. F., Coissac, E., Zundel, S., Riaz, T., Shehzad, W., Bessière, J., ... & Pompanon, F. (2010). An in silico approach for the evaluation of DNA barcodes. BMC Genomics, 11(1), 434.

Mercier, C., Boyer, F., Bonin, A., & Coissac, E. (2013, November). SUMATRA and SUMACLUST: fast and exact comparison and clustering of sequences. In Programs and Abstracts of the SeqBio 2013 workshop. Abstract (pp. 27-29).

Taberlet, P., Bonin, A., Zinger, L., & Coissac, E. (2018). Environmental DNA: For Biodiversity Research and Monitoring. Oxford University Press.

Zinger, L., Taberlet, P., Schimann, H., Bonin, A., Boyer, F., De Barba, M., ... & Chave, J. (2019). Body size determines soil community assembly in a tropical forest. Molecular Ecology, 28(3), 528-543.
}
\keyword{dataset}
