% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_subset_metabarlist.R
\name{subset_metabarlist}
\alias{subset_metabarlist}
\title{Subset a \code{metabarlist} object}
\usage{
subset_metabarlist(metabarlist, table, indices)
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object to be subsetted}

\item{table}{the table where the information on which the subsetting is based. Can be one of `reads`, `motus`, `pcrs`, or `samples`.}

\item{indices}{a boolean vector indicating the elements, i.e. rows or columns to keep. Elements to keep should be encoded \code{TRUE}.}
}
\value{
a \code{metabarlist} object similar to `metabarlist` except that it contain only the selected elements.
}
\description{
Subsetting a \code{\link{metabarlist}} object
}
\details{
Subsetting a \code{metabarlist} will select specific rows from the `reads`, `motus`, `pcrs`, or `samples` tables. Factor levels that are unused after selection are dropped. Note however that:
\itemize{
\item {If the selection is done on `reads`, `pcrs` or `samples`, the MOTUs not occurring in the pcrs or samples selected (i.e. those having a total number of reads of 0) are dropped.}
\item {If the selection is done on `motus`, the pcrs and samples where none of the selected MOTUs are found are kept.}
}
}
\examples{

data(soil_euk)


# Create a subset of soil_euk containing only annelid MOTUs
# (i.e. by searching for "Annelida" in MOTUs taxonomic assignements)
annelids <- subset_metabarlist(soil_euk,
  table = "motus",
  indices = grepl("Annelida", soil_euk$motus$path)
)
summary_metabarlist(annelids)

# Create a subset of soil_euk containing only pcrs conducted in plate 1
plate1 <- subset_metabarlist(soil_euk,
  table = "pcrs",
  indices = (soil_euk$pcrs$plate_no == 1)
)

summary_metabarlist(plate1)

# Create a subset of soil_euk containing only positive controls (i.e. excluding biological samples too who are NA in soil_euk$pcrs$control_type)
poscontrols <- subset_metabarlist(soil_euk,
  table = "pcrs",
  indices = (soil_euk$pcrs$control_type == "positive" & !is.na(soil_euk$pcrs$control_type))
)

summary_metabarlist(poscontrols)


}
\author{
Lucie Zinger
}
