% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_taxodecider.R
\name{taxodecider}
\alias{taxodecider}
\title{Selecting the good taxonomy}
\usage{
taxodecider(metabarlist, best.db, sim.scores, lineage, threshold)
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object}

\item{best.db}{a vector of two database names, the best one in terms of taxonomic information
reliability being provided first.}

\item{sim.scores}{a vector of two column names in the `motus` table corresponding to the similarity scores
of each database.}

\item{lineage}{a vector of two column names in the `motus` table corresponding to the full taxonomic
lineage obtained for each database. Should be provided in the same order than `best.db`.
\code{"centroid"} or \code{"pairwise"}. Default is \code{"centroid"}.}

\item{threshold}{a similiarty score threshold above which the annotation of the best database is selected if
both databases yields high similarity scores.}
}
\value{
a \code{metabarlist} object with a dataframe motus including the preferred taxonomic assignements
}
\description{
Selecting a given taxonomic annotation between those obtained from two different databases
}
\details{
The function \code{taxodecider} allows choosing amongst two taxonomic annotation based on the best identity scores and preferences for a given database (e.g. more reliable taxonomy). All taxonomic information should be stored in `metabarlist$motus`.
}
\examples{

\dontrun{
data(soil_euk)
soil_euk <- silva_annotator(
   metabarlist = soil_euk,
   silva.path = "~/Documents/workspace/metabaRffe_external_data/lit_euk---ssu---otus.csv",
   clust.path =  "~/Documents/workspace/metabaRffe_external_data/lit_euk---ssu---sequence_cluster_map---litiere_euk_cl97_agg_filt.clstr")

soil_euk$motus$similarity = soil_euk$motus$similarity/100

soil_euk2 <- taxodecider(
   metabarlist = soil_euk,
   best.db = c("silva", "embl"),
   sim.scores = c("similarity", "best_identity.order_filtered_embl_r136_noenv_EUK"),
   lineage = c("lineage_silva", "path"),
   threshold = 0.9
)

}

}
\author{
Anne-Sophie Benoiston, Lucie Zinger
}
