% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_motus.R
\name{aggregate_motus}
\alias{aggregate_motus}
\alias{FUN_agg_motus_sum}
\title{Aggregate MOTUs}
\usage{
aggregate_motus(metabarlist, groups = NULL, FUN = FUN_agg_motus_sum)

FUN_agg_motus_sum(metabarlist, groups)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}

\item{groups}{a grouping vector or factor for MOTUs.
\code{NA} values treated as a group level}

\item{FUN}{a function of MOTU aggregation.
Default is the sum of MOTUs reads abundances for each grouping value.}
}
\value{
A \code{metabarlist} where the table `reads` contains MOTUs abundances aggregated according to a grouping vector/factor (e.g. taxonomic assignment at the phylum level), using a method defined in \code{FUN} and where number of columns of tables `reads` will be equal to that of the number of groups in `groups`.
}
\description{
Aggregate MOTUs in a \code{metabarlist} object according to a grouping factor or vector
}
\details{
The function \code{aggregate_motus} is typically used for aggregating MOTUs at a given taxonomic resolution. The user is free to use its own method of aggregation, but the most common aggregation method is to sum reads for each taxa and is therefore pre-encoded in \code{FUN_agg_motus_sum}.

After aggregation, the information contained in the `motus` table corresponds to the
information of the most abundant MOTU in a given group.
}
\section{Functions}{
\itemize{
\item \code{aggregate_motus}: Aggregate MOTUs in a \code{metabarlist} object according to a grouping factor or vector for each pcr.

\item \code{FUN_agg_motus_sum}: compute the sum of reads according to a grouping factor or vector for each pcr.
}}

\examples{

data(soil_euk)

## With MOTU phylum assignment as grouping factor and
## default grouping aggregation (sum reads across replicates)
soil_euk_ag <- aggregate_motus(soil_euk, groups = soil_euk$motus$phylum_name)
summary_metabarlist(soil_euk)
summary_metabarlist(soil_euk_ag)

## With a custom function (here equivalent to FUN_agg_sum,
## i.e. summing all MOTUs abundance across groups)
soil_euk_ag <- aggregate_motus(soil_euk, groups = soil_euk$motus$phylum_name,
                               FUN = function(metabarlist, groups){
                                       t(rowsum(t(metabarlist$reads), groups))})

}
\seealso{
\code{\link{aggregate_pcrs}}, \code{\link{apply}}, \code{\link{aggregate}}
}
\author{
Lucie Zinger
}
