% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_check_metabarlist.R
\name{check_metabarlist}
\alias{check_metabarlist}
\title{Test that the provided list is a well formed metabaRffe list}
\usage{
check_metabarlist(metabarlist)
}
\arguments{
\item{metabarlist}{a \code{\link{TODEFINE}} object}
}
\value{
TRUE or throws an error
}
\description{
Test that the provided list is a well formed metabaRffe list
}
\details{
Check for the properties awaited for a well formed metabaRffe list:
\itemize{
\item {is a list with three attributes named `reads`, `motus` and `pcrs`}
\item {`reads` is a numeric matrix}
\item {`motus` and `pcrs` are data.frame}
\item {`motus` and `pcrs` have mandatory columns, i.e. tag_fwd, tag_rev, primer_fwd, primer_rev, plate_no, plate_col and plate_row for `pcrs`
  and sequence for `motus`}
}

Moreover, the function issues a warning if any sample or OTU is associated to a count of 0
}
\examples{

data(soil_euk)

check_metabarlist(soil_euk)

}
\author{
Clément Lionnet & Frédéric Boyer
}
