% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_check_metabarlist.R
\name{check_metabarlist}
\alias{check_metabarlist}
\title{Check that a list of tables is in the correct format as a metabarlist}
\usage{
check_metabarlist(metabarlist)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}
}
\value{
TRUE or throws a stop
}
\description{
Test that a list of tables containing information on MOTU abundances, MOTU characteristics, pcr characteristics, and sample characteristics form a congruent \code{\link{metabarlist}}
}
\details{
Checks for the properties required for a well formed \code{metabarlist} object:

\itemize{
\item {is a list with four attributes named `reads`, `motus`, `pcrs` and `samples`}
\item {`reads` is a numeric matrix}
\item {`motus`, `pcrs` and `samples` are of type `data.frame`}
\item {`reads` must have the exact same row names and columns as the row names of `motus` and `pcrs` (in the same order)}
\item {`pcrs` have mandatory columns, i.e.  `sample_id`, `type` and `control_type`}
\item {values in `type` are properly defined, i.e. `sample` or `control`}
\item {values in and `control_type` are properly defined, i.e. `sequencing`, `pcr`, `extraction`, `positive`}
}

The function will stop if these basic criteria are not met.

The function will issue warnings if tables lack columns which are non-mandatory for the \code{metabaRffe} package to run, but which are mandatory for specific functions (e.g., \code{ggpcrplate})
\itemize{
\item {the column `sequence` for the `motus` table}
\item {the columns `tag_fwd`, `tag_rev`, `primer_fwd`, `primer_rev`, `plate_no`, `plate_col`, and `plate_row` for the `pcrs` table}
}


In addition, the function issues a warning if any sample or MOTU is has a count of 0
}
\examples{

data(soil_euk)

check_metabarlist(soil_euk)

}
\author{
Clément Lionnet & Frédéric Boyer & Lucie Zinger
}
