% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_check_metabarlist.R
\name{check_metabarlist}
\alias{check_metabarlist}
\title{Checks if a list of tables is a metabarlist}
\usage{
check_metabarlist(metabarlist)
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object}
}
\value{
\code{check_metabarlist} returns TRUE if \code{metabarlist} fulfills of conditions to be of class \code{\link{metabarlist}}` (see details). If not, the function throws a stop.
}
\description{
Tests if a list of tables containing information on MOTUs abundances, MOTUs characteristics, PCRs characteristics, and samples characteristics form an object of class \code{\link{metabarlist}}
}
\details{
The function \code{check_metabarlist} checks if \code{metabarlist} has all the properties required to be a well formed \code{metabarlist} object:

\itemize{
\item {\code{metabarlist} should be is a list with four tables named
       `reads`, `motus`, `pcrs` and `samples`.}
\item {table `reads` should be a numeric matrix.}
\item {tables `motus`, `pcrs` and `samples` should be a data frame.}
\item {table `reads` row names should be exactely the same than row names in table `pcrs`.}
\item {table `reads` column names should be exactely the same than
       row names in table `motus`.}
\item {table `pcrs` should contain the mandatory columns,
       i.e.  `sample_id`, `type` and `control_type`.}
\item {values in `type` should be properly defined:
       they can only be `sample` or `control`.}
\item {values in and `control_type` should be properly defined:
       they can only be `sequencing`, `pcr`, `extraction`, `positive`, (or NA for samples).}
}

The function will stop if these basic criteria are not met.

\code{check_metabarlist} also issues warnings if tables lack columns which are non-mandatory for the \code{metabaRffe} package to run, but that are mandatory (or have a mandatory format) for specific functions (e.g., \code{\link{ggpcrplate}})
\itemize{
\item {the column `sequence` for the `motus` table}
\item {the columns `tag_fwd`, `tag_rev`, `primer_fwd`, `primer_rev`, `plate_no`, `plate_col`, and `plate_row` for the `pcrs` table}
}


In addition, the function issues a warning if any PCR or MOTU has a total count of 0 across the entire dataset.
}
\examples{

data(soil_euk)
check_metabarlist(soil_euk)

}
\seealso{
\code{\link{metabarlist_generator}}
}
\author{
Clément Lionnet & Frédéric Boyer & Lucie Zinger
}
