% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contaslayerrrr.R
\name{contaslayer}
\alias{contaslayer}
\title{Detects contaminants in metabarcoding data}
\usage{
contaslayer(x, controls, method = "all")
}
\arguments{
\item{x}{a \code{\link{TODEFINE}} object}

\item{controls}{a vector of amplicon names  corresponding to negative controls.}

\item{method}{a character string specifying the detection method to be used. Default is \code{"max"}}
}
\value{
a vector containing the names of OTUs identified as contaminants

#' @details
The assumption is that in negative controls, a contaminant should be preferentially amplified as there is no competing DNA.
#'  \code{method = "max"} returns OTU names for which frequencies across the entire dataset are maximum in at least one negative control
 \code{method = "all"} returns OTU names for which the average frequency across all negative controls is greater than across samples
}
\description{
Uses negative controls to detects contaminants OTUs in a \code{\link{TODEFINE}} object.
}
\author{
Lucie Zinger
}
