% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_contaslayer.R
\name{contaslayer}
\alias{contaslayer}
\title{Detecting contaminants in metabarcoding data using negative controls}
\usage{
contaslayer(metabarlist,
  controls = rownames(metabarlist$pcrs)[which(metabarlist$pcrs$control_type
  == "pcr")], method = "max")
}
\arguments{
\item{controls}{a vector of amplicon names corresponding to negative controls.}

\item{method}{a character string specifying the detection method to be used. Default is \code{"max"}}

\item{x}{a \code{\link{metabarlist}} object}
}
\value{
a vector containing the names of OTUs identified as contaminants
}
\description{
Uses negative controls to detect contaminant OTUs in a \code{\link{TODEFINE}} object.
}
\details{
In negative controls, a contaminant should be preferentially amplified since there is no competing DNA. \code{\link{contaslayer}} relies on this assumption and detects OTUs whose relative abundance across the whole dataset is maximum in negative controls.
\code{method = "max"} returns the names of OTUs whose frequencies across the entire dataset are maximum in at least one negative control
\code{method = "all"} returns the names of OTUs whose frequencies across all negative controls is greater than that across all samples
}
\examples{

data(soil_euk)

#finds contaminants from PCR amplification
contaminant = contaslayer(soil_euk)
head(soil_euk$motus[contaminant,])

#Distribution of the most abundant contaminants in the PCR plate design
max.conta = contaminant[which.max(soil_euk$motus[contaminant, "count"])]

p <- ggpcrplate(soil_euk, legend_title = "# reads",
           FUN = function(m){m$reads[,max.conta]})
p + scale_size(limits=c(1,max(soil_euk$reads[,max.conta]))) +
     ggtitle("Distribution of the most abundant contaminant")

}
\author{
Lucie Zinger
}
