% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_contaslayer.R
\name{contaslayer}
\alias{contaslayer}
\title{Detecting contaminants in metabarcoding data using negative controls}
\usage{
contaslayer(x, controls, method = "max")
}
\arguments{
\item{x}{a \code{\link{TODEFINE}} object}

\item{controls}{a vector of amplicon names corresponding to negative controls.}

\item{method}{a character string specifying the detection method to be used. Default is \code{"max"}}
}
\value{
a vector containing the names of OTUs identified as contaminants
}
\description{
Uses negative controls to detect contaminants OTUs in a \code{\link{TODEFINE}} object.
}
\details{
In negative controls, a contaminant should be preferentially amplified as there is no competing DNA. \code{\link{contaslayer}} relies on this assumption and detects OTUs of which relative abundance across the whole dataset is maximum in negative controls.
\code{method = "max"} returns OTU names for which frequencies across the entire dataset are maximum in at least one negative control
\code{method = "all"} returns OTU names for which the frequency across all negative controls is greater than that across all samples
}
\examples{

data(soil_euk)

#finds contaminants from PCR amplification
pcr.controls = rownames(soil_euk$pcrs)[which(soil_euk$pcrs$control_type=="pcr")]
contaminant = contaslayer(soil_euk$reads, controls = pcr.controls)
head(soil_euk$motus[contaminant,])

#Distribution of the most abundance contaminant in the PCR plate design
max.conta = contaminant[which.max(soil_euk$motus[contaminant, "count"])]

p <- ggpcrplate(soil_euk, legend_title = "# reads",
           FUN = function(m){m$reads[,max.conta]})
p + scale_size(limits=c(1,max(soil_euk$reads[,max.conta]))) +
     ggtitle("Distribution of the most abundant contaminant")

}
\author{
Lucie Zinger
}
