% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_contaslayer.R
\name{contaslayer}
\alias{contaslayer}
\title{Detecting contaminants in metabarcoding data using negative controls}
\usage{
contaslayer(x, controls, method = "max")
}
\arguments{
\item{x}{a \code{\link{TODEFINE}} object}

\item{controls}{a vector of amplicon names corresponding to negative controls.}

\item{method}{a character string specifying the detection method to be used. Default is \code{"max"}}
}
\value{
a vector containing the names of OTUs identified as contaminants
}
\description{
Uses negative controls to detect contaminant OTUs in a \code{\link{TODEFINE}} object.
}
\details{
In negative controls, a contaminant should be preferentially amplified since there is no competing DNA. \code{\link{contaslayer}} relies on this assumption and detects OTUs whose relative abundance across the whole dataset is maximum in negative controls.
\code{method = "max"} returns the names of OTUs whose frequencies across the entire dataset are maximum in at least one negative control
\code{method = "all"} returns the names of OTUs whose frequencies across all negative controls is greater than that across all samples
}
\examples{

data(soil_euk)

#finds contaminants from PCR amplification
pcr.controls = rownames(soil_euk$pcrs)[which(soil_euk$pcrs$Control_type=="PCR")]
contaminant = contaslayer(soil_euk$reads, controls = pcr.controls)
head(soil_euk$motus[contaminant,])

#Distribution of the most abundant contaminants in the PCR plate design
max.conta = contaminant[which.max(soil_euk$motus[contaminant, "count"])]
p = ggpcrplate(attr = soil_euk$reads[,max.conta], plate_no = soil_euk$pcrs$plate_no,
                  plate_col = soil_euk$pcrs$plate_col,
                  plate_row =  soil_euk$pcrs$plate_row,
                  control_type = soil_euk$pcrs$Control_type)
p + scale_size(limits=c(1,max(soil_euk$reads[,max.conta]))) +
     labs(size="# reads", fill="control type") +
     scale_fill_manual(values=c("brown", "pink", "cyan4", "red"),
                       na.value = "white") +
     ggtitle("Distribution of the most abundant contaminant")

}
\author{
Lucie Zinger
}
