% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_contaslayer.R
\name{contaslayer}
\alias{contaslayer}
\title{Detecting genuine MOTUs from contaminations using negative controls}
\usage{
contaslayer(metabarlist, method = "max", control_types = c("pcr",
  "extraction"), controls = NULL, output_col = "not_contamination")
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object}

\item{method}{a character string specifying the detection method to be used. Default: \code{"max"}}

\item{control_types}{a vector of control types contained in the column control_type of metabarlist$pcrs. This parameter is not used when the parameter \code{controls} is specified. Default is c("pcr", "extraction").}

\item{controls}{a vector of amplicon names corresponding to negative controls.}

\item{output_col}{a character string for the column name in `metabarlist$motus` in which the result will be stored. Default is "not_contamination"}
}
\value{
a metabarlist with a new boolean column vector of name `output_col` in `metabarlist$motus` indicating whether MOTUs are genuine (\code{TRUE}) or identified as contaminants (\code{FALSE})
}
\description{
Uses negative controls to determine if MOTUs in a \code{\link{metabarlist}} object are most likely genuine or most likely contaminants.
}
\details{
In negative controls, a contaminant should be preferentially amplified since there is no competing DNA. \code{\link{contaslayer}} relies on this assumption and detects MOTUs whose relative abundance across the whole dataset is maximum in negative controls.
\code{method = "max"} considers a MOTU as a contaminant if its frequencies across the entire dataset are maximum in at least one negative control
\code{method = "all"} considers a MOTU as a contaminant if its frequencies across all negative controls is greater than that across all samples
}
\examples{

data(soil_euk)

# identify potential genuine from contaminants using PCR or extraction negative controls
out <- contaslayer(soil_euk)
tail(colnames(out$motus))
head(out$motus[which(out$motus$not_contamination==F),])
length(which(out$motus$not_contamination==F))

# Distribution of the most abundant contaminants in the PCR plate design
max.conta <- rownames(out$motus)[out$motus$not_contamination==F][which.max(out$motus[contaminant, "count"])]

p <- ggpcrplate(soil_euk,
  legend_title = "# reads",
  FUN = function(m) {
    m$reads[, max.conta]
  }
)

p + scale_size(limits = c(1, max(soil_euk$reads[, max.conta]))) +
  ggtitle("Distribution of the most abundant contaminant")

# identify potential genuine from contaminants using extraction negative controls only
out <- contaslayer(soil_euk, control_types="extraction", output_col= "not_ext_contamination")
tail(colnames(out$motus))
length(which(out$motus$not_ext_contamination==F))

}
\author{
Lucie Zinger, Clement Lionnet
}
