% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_fasta_generator.R
\name{fasta_generator}
\alias{fasta_generator}
\title{Generate a fasta file on a subset of sequences}
\usage{
fasta_generator(file, id, seq)
}
\arguments{
\item{file}{path/name of the output}

\item{id}{a character vector containing the identifiants of the motus to export}

\item{seq}{a character vector containing the sequence corresponding to the motus id to export}
}
\value{
a fasta file
}
\description{
Generate a fasta file on a subset of sequences from a \code{\link{TODEFINE}} object.
}
\details{
Creates a fasta file from a set of sequences of interest. The header can be easily amended with further informations with e.g. \code{paste}
}
\examples{

data(soil_euk)

#export in fasta the 10 most abundant MOTUs
idx = order(soil_euk$motus$count, decreasing=T)[1:10]
fasta_generator("Dominants.fasta", rownames(soil_euk$motus)[idx],
                soil_euk$motus$sequence[idx])

#export in fasta the 10 most abundant MOTUs and their abundance
idx = order(soil_euk$motus$count, decreasing=T)[1:10]
headers = paste(rownames(soil_euk$motus)[idx],
                "abundance=", soil_euk$motus$count[idx],
                sep="")
fasta_generator("Dominants.fasta", headers,
                soil_euk$motus$sequence[idx])
}
\author{
Lucie Zinger
}
