% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_ggpcrplate.R
\name{ggpcrplate}
\alias{ggpcrplate}
\title{Plotting a PCR attribute in the PCR plate design}
\usage{
ggpcrplate(metabarlist, table, index)
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object}

\item{table}{the table where the information on which the plotting is based. Can be only `reads`,  or `pcrs`.}

\item{index}{the name of the column name containing the information to be plotted. This information should be a numeric vector.}
}
\value{
a ggplot
}
\description{
Plots an attribute of PCRs from a \code{\link{metabarlist}} object in the correspondint 96-well PCR plates design
}
\details{
Vizualizing some attributes of the PCRs in their plate design context, i.e. their location in one or a set of 96-well PCR plates. This can be useful to identify potential problems (e.g. low amount of reads due to non-functional primer sets).
}
\examples{

data(soil_euk)

#Plot the number of reads per pcrs

##Store the number of reads per pcrs in the pcrs table
soil_euk$pcrs$seq_depth = rowSums(soil_euk$reads)

##Plot the results
ggpcrplate(soil_euk, table = "pcrs", index = "seq_depth")


#Plot the number of reads of the most abundant MOTU

##Get the name of the most abundant MOTU
id = colnames(soil_euk$reads)[which.max(colSums(soil_euk$reads))]

##Plot the results
ggpcrplate(soil_euk, table = "reads", index = id) + labs(size="#reads of most\\nabundant MOTU")


}
\author{
Lucie Zinger
}
