% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_ggpcrplate.R
\name{ggpcrplate}
\alias{ggpcrplate}
\title{Plotting a PCR attribute in the PCR plate design}
\usage{
ggpcrplate(metabarlist, legend_title = "well_values",
  FUN = function(metabarlist) {     rowSums(metabarlist$reads) })
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object.}

<<<<<<< HEAD
\item{legend_title}{the title of legend containing the plotted information.}
=======
\item{table}{the table containing the information used for plotting. Can be either `reads`,  or `pcrs`.}
>>>>>>> 55486647798b8c099130fd3e5c1e38c0dee2d2f7

\item{FUN}{a function which return a vector containing the information to be plotted. The vector should be a numeric vector which has the same length of table 'reads'.}
}
\value{
a ggplot
}
\description{
Plots an attribute of PCRs from a \code{\link{metabarlist}} object into the corresponding 96-well PCR plate design
}
\details{
Vizualizing some attributes of the PCRs in their plate design context, i.e. their location in one or a set of 96-well PCR plates. This can be useful to identify potential problems (e.g. high amount of reads in one control due to cross contaminations with neighboring samples).
}
\examples{

data(soil_euk)

<<<<<<< HEAD
#Plot the number of reads per pcrs
ggpcrplate(soil_euk)
=======
#Plot the number of reads per pcr

##Store the number of reads per pcr in the pcrs table
soil_euk$pcrs$seq_depth = rowSums(soil_euk$reads)

##Plot the results
ggpcrplate(soil_euk, table = "pcrs", index = "seq_depth")
>>>>>>> 55486647798b8c099130fd3e5c1e38c0dee2d2f7


#Plot the number of reads of the most abundant MOTU
ggpcrplate(soil_euk, legend_title = "#reads of most\\nabundant MOTU",
           FUN = function(m){m$reads[,which.max(colSums(m$reads))]})


}
\author{
Lucie Zinger
}
