% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpcrplate.R
\name{ggpcrplate}
\alias{ggpcrplate}
\title{Plotting a PCR attribute in the PCR plate design}
\usage{
ggpcrplate(
  metabarlist,
  legend_title = "well_values",
  FUN = function(metabarlist) {     rowSums(metabarlist$reads) }
)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object.}

\item{legend_title}{the title of legend containing the plotted information.}

\item{FUN}{a function which returns a vector containing the information to be plotted.
The vector should be numeric and of equal length to the number of
rows of the `reads` table.}
}
\value{
a ggplot
}
\description{
Plots an attribute of PCRs from a \code{metabarlist} object into the corresponding 96-well PCR plate design
}
\details{
Visualising PCR attributes in their plate design context, i.e. their location in one or a set of 96-well PCR plates. This can be useful for identifying potential problems (e.g. high amount of reads in one control due to cross contaminations with neighboring samples).
}
\examples{

data(soil_euk)

## Plot the number of reads per pcr
ggpcrplate(soil_euk)


## Plot the number of reads of the most abundant MOTU
ggpcrplate(soil_euk,
  legend_title = "#reads of most \nabundant MOTU",
  FUN = function(m) {
    m$reads[, which.max(colSums(m$reads))]
  }
)
}
\seealso{
\code{\link{ggpcrtag}}
}
\author{
Lucie Zinger, Clément Lionnet
}
