% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_ggpcrtags.R
\name{ggpcrtag}
\alias{ggpcrtag}
\title{Plotting a PCR attribute in the tag pair design}
\usage{
ggpcrtag(
  metabarlist,
  legend_title = "well_values",
  FUN = function(metabarlist) {     rowSums(metabarlist$reads) },
  taglist = as.character(unique(metabarlist$pcrs$tag_rev))
)
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object}

\item{legend_title}{the title of legend containing the plotted information.}

\item{FUN}{a function which return a vector containing the information to be plotted. The vector should be a numeric vector which has the same length of table 'reads'.}

\item{taglist}{a character vector corresponding to the full list of tags ordered as they are used in the PCR plate scheme}
}
\value{
a ggplot
}
\description{
Plots an attribute of PCRs from a \code{\link{metabarlist}} object in the tag pair design
}
\details{
Vizualizing attributes of the PCRs in their tag pair design context, i.e. according to the tag combination used. This can be useful to identify potential problems (e.g. low amount of reads due to non-functional primer/tag sets).
}
\examples{

data(soil_euk)

# Plot the number of reads per pcrs
ggpcrtag(soil_euk)

# Plot the number of reads of the most abundant MOTU
ggpcrtag(soil_euk,
  legend_title = "#reads of most\nabundant MOTU",
  FUN = function(m) {
    m$reads[,which.max(colSums(m$reads))]
  }
)
}
\author{
Lucie Zinger
}
