% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_ggplate.R
\name{ggplate}
\alias{ggplate}
\title{Plotting an attribute in the PCR plate design}
\usage{
ggplate(attr, plate_no, plate_col, plate_row, type)
}
\arguments{
\item{x}{a \code{\link{TODEFINE}} object}
}
\value{
a ggplot
}
\description{
Plots an attribute of amplicons from a \code{\link{TODEFINE}} object in the correspondint PCR plate design
}
\details{
TO WRITE
}
\examples{

data(soil_euk)
#plot number of reads per pcrs
seqdepth = rowSums(soil_euk$reads)
ggplate(attr = seqdepth, plate_no = soil_euk$pcrs$plate_no,
                  plate_col = soil_euk$pcrs$plate_col,
                  plate_row =  soil_euk$pcrs$plate_row,
                  control_type = soil_euk$pcrs$Control_type)

}
\author{
Fred Boyer, Lucie Zinger
}
