% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabarlist_to_phyloseq.R
\name{metabarlist_to_phyloseq}
\alias{metabarlist_to_phyloseq}
\title{Converts a metabarlist object in a phyloseq object}
\usage{
metabarlist_to_phyloseq(metabarlist, tax_colnames)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}

\item{tax_colnames}{a character vector with the column names to keep from the \code{motus} table of the \code{metabarlist} object}
}
\value{
A \code{phyloseq} object with three components: \code{otu_table}, \code{sample_data} and \code{taxonomy_table}
}
\description{
Converts a \code{metabarlist} object in a \code{phyloseq} object for downstream diversity analyses.
}
\details{
The function \code{metabarlist_to_phyloseq} converts a \code{metabarlist} object into a \code{phyloseq} object. A \code{phyloseq} object can contain several component data, such as \code{otu_table}, \code{sample_data} and \code{taxonomy_table}. The \code{otu_table} is equivalent to the \code{reads} matrix, the \code{sample_data} is equivalent to the \code{samples} data frame and the \code{taxonomy_table} corresponds to the taxonomic informations that may be contained in the \code{motus} data frame.
The sample names in the \code{reads} matrix (i.e. rownames) should be the the same as in the \code{samples} data frame.
}
\examples{

data(soil_euk) # load soil_euk data
soil_euk <- subset_metabarlist(soil_euk, "pcrs", metabarlist$pcrs$type == "sample") # select only samples, i.e. delete controls
soil_euk_agg <- aggregate_pcrs(soil_euk) # aggregate pcrs at sample level
tax_colnames <- c("phylum_name", "class_name", "order_name", "family_name", "genus_name", "species_name") # create vector with column names to keep from the motus table
soil_euk_physeq <- metabarlist_to_phyloseq(soil_euk, tax_colnames)
soil_euk_physeq

}
\seealso{
the \code{phyloseq} package
}
\author{
Anne-Sophie Benoiston
}
