% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_read_ngsfilter.R
\name{read_ngsfilter}
\alias{read_ngsfilter}
\title{Import a ngsfilter file}
\usage{
read_ngsfilter(file, additional.sep = "=", ...)
}
\arguments{
\item{file}{path/name of the input}

\item{additional.sep}{character indicating the field separator in the additional info colum in ngsfilter}

\item{...}{other arguments to be pasted from \code{read.table}}
}
\value{
a table
}
\description{
Import a ngsfilter file formatted for the \code{OBITools} package.
}
\details{
Creates a sample table from a ngsfilter file formatted for the \code{OBITools} package.
}
\examples{

}
\author{
Lucie Zinger
}
