% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silva_annotator.R
\name{silva_annotator}
\alias{silva_annotator}
\title{Including SILVAngs pipeline taxonomic annotations}
\usage{
silva_annotator(metabarlist, silva.path, clust.path)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}

\item{silva.path}{path to a table from the SILVAngs pipeline,
typically zipfile > ssu > exports > xxx---ssu---otus.csv}

\item{clust.path}{path to a file from the SilvAngs pipeline indicating MOTU cluster membership,
typically zipfile > ssu > stats > sequence_cluster_map > data >
xxx---ssu---sequence_cluster_map---tmptaxo.clstr}
}
\value{
a \code{metabarlist} object with table `motus` including the taxonomic assignments from silva
}
\description{
Importing and formatting taxonomic annotations obtained with the SILVAngs pipeline (https://ngs.arb-silva.de/silvangs/) and including it in a \code{\link{metabarlist}} object.
}
\details{
Users can be interested in using the SILVAngs pipeline for assigning a taxon to DNA sequences/ MOTUs. Assuming that it is done on data already filtered (i.e. dereplicated, clustering etc.), resulting in one sequence per sequence or MOTU, one can then use the SILVAngs pipeline by setting all filtering parameters to "null" (i.e. returning to no filtration) and the taxonomic assignments parameters by default (or following the user's preferences). These taxonomic assignments are compiled in the data archives provided by SILVAngs, in which two files are important for the function `silva_annotator`:
\itemize{
\item{}{`zipfile>ssu>exports>xxx---ssu---otus.csv`: a csv file containing the taxonomic assignment for each OTU}
\item{}{`zipfile>ssu>stats>sequence_cluster_map>data>xxx---ssu---sequence_cluster_map---tmptaxo.clstr`: a file containing the mapping informations of sequences to their respective MOTUs. Even though the clustering is done before using the SILVAngs pipeline and that the clustering parameters in this pipeline are settled to 100\% identity of clustering, SILVAngs uses CDHit, which can group together OTUs sharing the same prefix/suffix. So it is common to retrieve less assignations than what was expected.}
}
}
\examples{

data(soil_euk)
soil_euk <- silva_annotator(
   metabarlist = soil_euk,
   silva.path = system.file("extdata", "lit_euk---ssu---otus.csv", package = "metabaR"),
   clust.path = system.file("extdata",
               "lit_euk---ssu---sequence_cluster_map---litiere_euk_cl97_agg_filt.clstr",
                package = "metabaR"))

ggplot(soil_euk$motus, aes(x=factor(1), fill=phylum_silva)) +
  geom_bar() + coord_polar("y") +
  theme_minimal() + labs(x=NULL, y=NULL) +
  theme(legend.position = "bottom")

}
\seealso{
\code{\link{taxodecider}}
}
\author{
Lucie Zinger, Anne-Sophie Benoiston
}
