% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_subset_metabarlist.R
\name{subset_metabarlist}
\alias{subset_metabarlist}
\title{Subset a \code{metabarlist} object}
\usage{
subset_metabarlist(metabarlist, table, indices)
}
\arguments{
\item{metabarlist}{a \code{\link{metabarlist}} object to be subsetted}

\item{table}{the table where the information on which the subsetting is based. Can be only `reads`, `motus`, `pcrs`, or `samples`.}

\item{indices}{a numeric vector of indices indicating the elements, rows or columns to keep in the selected table.}
}
\value{
a \code{metabarlist} object that contains only the selected elements.
}
\description{
Subsetting a \code{\link{metabarlist}} object
}
\details{
The subsetting will select particular rows from each table. Factor levels that are unused after selection are dropped. If the selection is done on `reads`, `pcrs` or `samples`, the MOTUs not occurring in the selection are dropped too. If the selection is done on `motus`, the pcrs and samples where none of the selected MOTUs are found are kept.
}
\examples{

data(soil_euk)


#create a subset of soil_euk containing only annelids MOTUs

annelids = subset_metabarlist(metabarlist, table = "motus",
                              indices = grep("Annelida", metabarlist$motus$path))
dim(annelids$motus)
dim(annelids$reads)

#create a subset of soil_euk containing only pcrs conducted in plate 1

plate1 = subset_metabarlist(metabarlist, table = "pcrs",
                            indices = which(metabarlist$pcrs$plate_no == 1))

dim(plate1$reads)
dim(plate1$motus)

}
\author{
Lucie Zinger
}
