% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metabarlist.R
\name{summary_metabarlist}
\alias{summary_metabarlist}
\title{Metabarlist summary}
\usage{
summary_metabarlist(metabarlist, method = "dataset", groups = NULL)
}
\arguments{
\item{metabarlist}{a \code{metabarlist} object}

\item{method}{type of summary to provide. Should match with `dataset`, `motus` or `pcrs`.
Default is `dataset`}

\item{groups}{a grouping vector or factor of same number of rows than `motus` or `pcrs`
for which the summary should be done. Default is NULL}
}
\value{
The \code{summary_metabarlist} function returns basic summary statistics (nb of elements, reads and MOTUs in total or on average per pcrs) of a \code{metabarlist} object. The format of the value returned depends on the method used.
}
\description{
summary_metabarlist is a generic function used to produce summary statistics of a \code{metabarlist} object.
}
\details{
\code{summary_metabarlist} returns basic summary statistics of a \code{metabarlist} object. The summary returned depends on the `method` used :
\itemize{
\item{`dataset`}{returns a list of two data.frames: dataset_dimension contains the dimensions of the full metabarlist object, dataset_statistics contains the number of reads, motus in pcrs and samples, as well as average and sd values of reads and motus per sample}
\item{`motus` or `pcrs`}{returns a data.frame similar to the dataset_statistics described above according to a grouping factor/vector for MOTUs or pcrs}
}
}
\examples{

data(soil_euk)

## Dataset summary
summary_metabarlist(soil_euk, method = "dataset")

## Data summary per control type (NA = samples)
summary_metabarlist(soil_euk, method = "pcrs",
    groups = soil_euk$pcrs$control_type)

## Data summary per phyla
summary_metabarlist(soil_euk, method = "motus",
    groups = soil_euk$motus$phylum_name)

## Data summary per Habitat (i.e. to get from soil_euk$samples).
# Here, NA values correspond to technical controls
summary_metabarlist(soil_euk, method = "pcrs",
    groups = soil_euk$samples$Habitat[match(soil_euk$pcrs$sample_id, rownames(soil_euk$samples))])

}
\seealso{
\code{\link{summary}}
}
\author{
Lucie Zinger
}
