% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_tagjumpslayer.R
\name{tagjumpslayer}
\alias{tagjumpslayer}
\title{Filtering potential tag jumps in metabarcoding data}
\usage{
tagjumpslayer(x, threshold = 0.03)
}
\arguments{
\item{x}{a \code{\link{TODEFINE}} object}

\item{threshold}{an OTU relative abundance value in an amplicon below which it is considered to be absence}
}
\value{
a transformed OTU count matrix
}
\description{
Sets to zero abundances of potential tagjumps in a \code{\link{TODEFINE}} object.
}
\details{
TO WRITE, cite carlsen, esling, etc..
}
\examples{

data(soil_euk)
soil_euk_clean = tagjumpslayer(soil_euk$reads, 0.03)
range(colSums(soil_euk$reads))
range(colSums(soil_euk_clean))
#to improve & add verif plot

}
\author{
Lucie Zinger
}
